Create role.ESBMessaging.send.png
This screenshot shows the role that must be assigned to a developer in the subaccount of the tenant when the developer's credentials are used in the sender system to get authorization to call a CPI Soap service defined in an integration process and process messages within that process.  Further documentation on this is available in the issue list, and the documentation found as .md files in this repository.